+++ drivers/thermal/backward_compatible.c
@@ -1,3 +1,16 @@
+/*
+ * Copyright (C) 2015 MediaTek Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ */
+
 #define pr_fmt(fmt) KBUILD_MODNAME ": " fmt
 
 #include <linux/module.h>
@@ -9,19 +22,19 @@
  * backward_compatible_throttle
  * @tz - thermal_zone_device
  *
- * This function update the cooler state by monitoring the current temperature and trip points
+ * This function update the cooler state by monitoring the current
+ * temperature and trip points
  */
-static int backward_compatible_throttle(struct thermal_zone_device *tz, int trip)
+static int backward_compatible_throttle(struct thermal_zone_device *tz,
+	 int trip)
 {
-	long trip_temp;
+	int trip_temp;
 	struct thermal_instance *instance;
 
 	if (trip == THERMAL_TRIPS_NONE)
 		trip_temp = tz->forced_passive;
 	else
 		tz->ops->get_trip_temp(tz, trip, &trip_temp);
-
-	/* mutex_lock(&tz->lock); */
 
 	list_for_each_entry(instance, &tz->thermal_instances, tz_node) {
 		if (instance->trip != trip)
@@ -35,15 +48,12 @@ static int backward_compatible_throttle(struct thermal_zone_device *tz, int trip
 		thermal_cdev_update(instance->cdev);
 	}
 
-	/* mutex_unlock(&tz->lock); */
-
 	return 0;
 }
 
 static struct thermal_governor thermal_gov_backward_compatible = {
 	.name = "backward_compatible",
 	.throttle = backward_compatible_throttle,
-	/* .owner                = THIS_MODULE, */
 };
 
 static int __init thermal_gov_backward_compatible_init(void)
